+++ bin/.gitignore
@@ -2,4 +2,3 @@
 /Clases/
 /Controles/
 /Vistas/
-/Estilos/
