--- bin/Vistas/gerente_actualizar_personal1.fxml
@@ -1,24 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<?import javafx.scene.control.*?>
-<?import com.jfoenix.controls.*?>
-<?import javafx.scene.text.*?>
-<?import java.lang.*?>
-<?import javafx.scene.layout.*?>
-<?import javafx.scene.layout.AnchorPane?>
-
-<AnchorPane prefHeight="635.0" prefWidth="700.0" style="-fx-background-color: #ffffff;" xmlns="http://javafx.com/javafx/8" xmlns:fx="http://javafx.com/fxml/1" fx:controller="Controles.ControlActPersonal1">
-   <children>
-      <JFXTextField fx:id="idEntrada" focusColor="#808080" labelFloat="true" layoutX="184.0" layoutY="254.0" prefHeight="30.0" prefWidth="333.0" promptText="Ingrese la cédula del usuario" style="-fx-prompt-text-fill: #bbb;" unFocusColor="#ff7070">
-         <font>
-            <Font size="16.0" />
-         </font>
-      </JFXTextField>
-      <JFXButton fx:id="buscar" layoutX="274.0" layoutY="335.0" onAction="#cargarInterfazAP2" prefHeight="47.0" prefWidth="153.0" style="-fx-background-color: #ff7070;" text="Buscar" textFill="WHITE">
-         <font>
-            <Font size="16.0" />
-         </font>
-      </JFXButton>
-      <Label fx:id="errorID" layoutX="184.0" layoutY="299.0" prefHeight="17.0" prefWidth="333.0" textFill="#ff7070" />
-   </children>
-</AnchorPane>
