+++ bin/Vistas/gerente_registro_items.fxml
@@ -42,7 +42,7 @@
                   <Font size="16.0" />
                </font>
             </JFXTextField>
-            <JFXComboBox fx:id="comboBoxSede" disable="true" editable="true" layoutX="46.0" layoutY="273.0" onMouseClicked="#onClickedSedes" prefHeight="25.0" prefWidth="234.0" promptText="Seleccione sede" styleClass= "comboBoxGerente" />
+            <JFXComboBox fx:id="comboBoxSede" disable="true" layoutX="46.0" layoutY="273.0" onMouseClicked="#onClickedSedes" prefHeight="25.0" prefWidth="234.0" promptText="Seleccione sede" styleClass= "comboBoxGerente" />
             <Label layoutX="46.0" layoutY="246.0" prefHeight="25.0" prefWidth="100.0" text="Numero sede" textFill="#aeaeae">
                <font>
                   <Font size="16.0" />
