+++ src/Controles/ControlAdminConsultarYActualizar.java
@@ -10,14 +10,27 @@ import javafx.collections.FXCollections;
 import javafx.collections.ObservableList;
 import javafx.collections.transformation.FilteredList;
 import javafx.collections.transformation.SortedList;
+import javafx.event.ActionEvent;
 import javafx.fxml.FXML;
+import javafx.fxml.FXMLLoader;
+import javafx.scene.Parent;
+import javafx.scene.Scene;
+import javafx.scene.control.Button;
+import javafx.scene.control.Label;
 import javafx.scene.control.TableColumn;
 import javafx.scene.control.TableView;
 import javafx.scene.control.cell.PropertyValueFactory;
+import javafx.scene.layout.Pane;
+import javafx.util.Duration;
 
+import java.io.IOException;
 import java.util.function.Function;
 import java.util.function.Predicate;
 
+import javafx.animation.Interpolator;
+import javafx.animation.KeyFrame;
+import javafx.animation.KeyValue;
+import javafx.animation.Timeline;
 import javafx.application.Application;
 import javafx.beans.binding.Bindings;
 import javafx.beans.property.ObjectProperty;
@@ -33,42 +46,45 @@ public class ControlAdminConsultarYActualizar {
 
     @FXML
     private TableView<Empleado> tablaIndiceEmpleados;
+    
+    @FXML
+    private JFXButton editar;
 
     @FXML
-    private JFXTextField campoCorreo;
+    private Label labelCorreo;
 
     @FXML
     private JFXButton botonActualizarGerente;
 
     @FXML
-    private JFXTextField campoTelefono;
+    private Label labelTelefono;
 
     @FXML
     private TableColumn<Empleado, String> columnaIdentificacion;
 
     @FXML
-    private JFXTextField campoGenero;
+    private Label labelGenero;
 
     @FXML
-    private JFXTextField campoId;
+    private Label labelId;
 
     @FXML
-    private JFXTextField campoSede;
+    private Label labelSede;
 
     @FXML
     private TableColumn<Empleado, String> columnaNombres;
 
     @FXML
-    private JFXTextField campoEstadoCivil;
+    private Label labelEstadoCivil;
 
     @FXML
-    private JFXTextField campoEstado;
+    private Label labelEstado;
 
     @FXML
-    private JFXTextField campoPerfil;
+    private Label labelPerfil;
 
     @FXML
-    private JFXTextField campoNombre;
+    private Label labelNombre;
     
     @FXML
     private JFXTextField buscarNombre;
@@ -124,7 +140,7 @@ public class ControlAdminConsultarYActualizar {
         empleado -> empleado.getId().toLowerCase().contains(buscarId.getText().toLowerCase()), 
         buscarId.textProperty()));
         
-        FilteredList<Empleado> filteredItems = new FilteredList<>(gerente.consultarGerentes(), p -> true);
+        FilteredList<Empleado> filteredItems = new FilteredList<>(gerente.consultarEmpleados(true), p -> true);
         tablaIndiceEmpleados.setItems(filteredItems);
 
         filteredItems.predicateProperty().bind(Bindings.createObjectBinding(
@@ -148,37 +164,81 @@ public class ControlAdminConsultarYActualizar {
     }
     
     public void inicializarEmpleados() {
-    	empleados = gerente.consultarGerentes();
+    	empleados = gerente.consultarEmpleados(true);
     	tablaIndiceEmpleados.setItems(empleados);
     }
 
     private void refrescarEtiquetas(Empleado empleado) {
         if (empleado != null) {
             // Fill the labels with info from the person object.
-            campoNombre.setText(empleado.getNombre());
-            campoId.setText(empleado.getId());
-            campoTelefono.setText(empleado.getTelefono());
-            campoEstado.setText(empleado.getEstado());
-            campoSede.setText(empleado.getSede());
-            campoCorreo.setText(empleado.getCorreo());
-            campoPerfil.setText(empleado.getPerfil());
-            campoEstadoCivil.setText(empleado.getEstadoCivil());
-            campoGenero.setText(empleado.getGenero());
+        	labelNombre.setText(empleado.getNombre());
+        	labelId.setText(empleado.getId());
+        	labelTelefono.setText(empleado.getTelefono());
+        	labelEstado.setText(empleado.getEstado());
+        	labelSede.setText(empleado.getSede());
+        	labelCorreo.setText(empleado.getCorreo());
+        	labelPerfil.setText(empleado.getPerfil());
+        	labelEstadoCivil.setText(empleado.getEstadoCivil());
+        	labelGenero.setText(empleado.getGenero());
 
             // TODO: We need a way to convert the birthday into a String! 
             // birthdayLabel.setText(...);
         } else {
             // Person is null, remove all the text.
-        	campoNombre.setText("");
-            campoId.setText("");
-            campoTelefono.setText("");
-            campoEstado.setText("");
-            campoSede.setText("");
-            campoCorreo.setText("");
-            campoPerfil.setText("");
-            campoEstadoCivil.setText("");
-            campoGenero.setText("");
+        	labelNombre.setText("");
+        	labelId.setText("");
+        	labelTelefono.setText("");
+        	labelEstado.setText("");
+            labelSede.setText("");
+            labelCorreo.setText("");
+            labelPerfil.setText("");
+            labelEstadoCivil.setText("");
+            labelGenero.setText("");
         }
+    }
+    
+    //Aplica el efecto de transición:
+    public void efectoCambio(FXMLLoader cargador, Empleado E, Pane panelCentral) throws IOException {
+		cargador.setLocation(getClass().getResource("/Vistas/gerente_actualizar_personal.fxml"));
+		Parent GUI = (Parent)cargador.load();    	    		
+		panelCentral.getChildren().clear();
+		panelCentral.getChildren().add(GUI);
+		Scene scene = GUI.getScene();
+		GUI.translateXProperty().set(scene.getWidth());
+		Timeline timeline = new Timeline();
+		KeyValue rango = new KeyValue(GUI.translateXProperty(), 0, Interpolator.EASE_IN);
+		KeyFrame duracion = new KeyFrame(Duration.seconds(0.4), rango);
+		timeline.getKeyFrames().add(duracion);
+		timeline.play();
+		ControlActPersonal controlador = cargador.getController();
+		controlador.cargar(E, true);
+    }
+    
+    //Extrae el empleado seleccionado:
+    public Empleado extraerEmpleado() {
+    	
+    	 String nombre = labelNombre.getText();
+         String id = labelId.getText();
+         String telefono = labelTelefono.getText();
+         String estado = labelEstado.getText();
+         String sede = labelSede.getText();
+         String correo = labelCorreo.getText();
+         String perfil = labelPerfil.getText();
+         String estadoCivil = labelEstadoCivil.getText();
+         String genero = labelGenero.getText();  
+         
+         Empleado E = new Empleado(nombre,id,telefono,estado,sede,correo,perfil,estadoCivil,genero);
+         
+         return E;
+         
+    }
+          
+    //Cambia de la interfaz de actualización de personal:
+    @FXML void cargarInterfazAP(ActionEvent event) throws IOException {
+    	FXMLLoader cargador = new FXMLLoader();   	    		
+		Pane panelCentral = (Pane)((Button)event.getSource()).getParent();
+		Empleado E = extraerEmpleado();
+		efectoCambio(cargador,E,panelCentral);
     }
 
 }
