+++ src/Vistas/administrador_GUI.fxml
@@ -50,7 +50,7 @@
                	<Button fx:id="atras" layoutX="12.0" layoutY="1.0" mnemonicParsing="false" onAction="#retroceder" opacity="0.6" prefHeight="25.0" prefWidth="25.0" style="-fx-graphic: url('/Imagenes/atras.png');" styleClass="controles" />                  
                   <Button fx:id="cerrar" layoutX="667.0" layoutY="2.0" mnemonicParsing="false" onAction="#cerrarAplicacion" prefHeight="25.0" prefWidth="25.0" style="-fx-graphic: url('/Imagenes/close.png');" styleClass="controles" />
                   <Button fx:id="minimizar" layoutX="632.0" layoutY="2.0" mnemonicParsing="false" onAction="#minimizarAplicacion" prefHeight="25.0" prefWidth="25.0" style="-fx-graphic: url('/Imagenes/minimize.png');" styleClass="controles" />
-                  <Text fx:id="titulo" fill="#ff7070" layoutX="90.0" layoutY="24.0" strokeType="OUTSIDE" strokeWidth="0.0" text="Actualizar Sedes" textAlignment="CENTER" wrappingWidth="519.13671875">
+                  <Text fx:id="titulo" fill="#ff7070" layoutX="90.0" layoutY="22.0" strokeType="OUTSIDE" strokeWidth="0.0" textAlignment="CENTER" wrappingWidth="519.13671875">
                      <font>
                         <Font size="17.0" />
                      </font>
