+++ src/Vistas/gerente_GUI.fxml
@@ -28,10 +28,9 @@
                   <Font size="16.0" />
                </font>
             </Text>
-            <JFXButton fx:id="botonActualizarPersonal" layoutX="43.0" layoutY="315.0" onAction="#cargarInterfazAP1" prefHeight="23.0" prefWidth="215.0" styleClass="botonesPanel" text="Actualizar Personal" textFill="GREY" />
-            <JFXButton fx:id="botonActualizaritems" layoutX="43.0" layoutY="386.0" onAction="#actualizarItems" prefHeight="23.0" prefWidth="215.0" styleClass="botonesPanel" text="Actualizar Items" textFill="GREY" />
-            <JFXButton fx:id="botonConsultarPersonal" layoutX="43.0" layoutY="424.0" onAction="#consultarPersonal" prefHeight="23.0" prefWidth="215.0" styleClass="botonesPanel" text="Consultar Personal" textFill="GREY" />
-            <JFXButton fx:id="botonConsultarSedes" layoutX="43.0" layoutY="460.0" onAction="#consultarSedes" prefHeight="23.0" prefWidth="215.0" styleClass="botonesPanel" text="Consultar Sedes" textFill="GREY" />
+            <JFXButton fx:id="botonActualizaritems" layoutX="43.0" layoutY="318.0" onAction="#actualizarItems" prefHeight="23.0" prefWidth="215.0" styleClass="botonesPanel" text="Actualizar Items" textFill="GREY" />
+            <JFXButton fx:id="botonConsultarPersonal" layoutX="43.0" layoutY="356.0" onAction="#consultarPersonal" prefHeight="23.0" prefWidth="215.0" styleClass="botonesPanel" text="Consultar Personal" textFill="GREY" />
+            <JFXButton fx:id="botonConsultarSedes" layoutX="43.0" layoutY="392.0" onAction="#consultarSedes" prefHeight="23.0" prefWidth="215.0" styleClass="botonesPanel" text="Consultar Sedes" textFill="GREY" />
             <ImageView fitHeight="15.0" fitWidth="18.0" layoutX="97.0" layoutY="597.0" pickOnBounds="true" preserveRatio="true">
                <image>
                   <Image url="@../Imagenes/cerrarSesion.png" />
